Add initial parser
This commit has a function for parsing a single line of source code to
an instruction and an argument (with the appropriate types). It
engendered a change in the spec: I added an argument to the `HALT`
instruction in order to keep the return type of the parsing function
uniform. I could ignore it or use it as the return code of the program.