Integrate the IO module
This commit integrates the IO module with the divarema engine. The
engine now has two type parameters, an input and output (that must
implement BufRead and Write) and the constructor takes two additional
args of those types.

Tests that don't use IO use the empty() and sink() functions from
std::io to create a io module.